Correct sdk_version
This change corrects sdk_version for libraries. Before the sdk_version
is not set, and by default the library will be built against the
implementation of the framework. These two libraries do not need to
build against the implemenation of the framework. Set the sdk_version to
the right scope.

Test: presubmit
Bug: 321743530
Change-Id: I6b5734e66c869c3091ed7e5b6ead458bc1ee78c2