Synchronize automatic disconnect and manually bind
(Manually copied over from google3.)

A scheduled automatic disconnection may be running interleaving with a
manuallyBind call. This is problematic because the disconnection in
rare cases destroys the latch of the current binding attempt.
Then, the binding will happen but the profile sender class gives up
on it straight after initializing. This results in a strange state,
where the system thinks we're bound and connected, but the connected
apps sdk has no reference to that connection.

Bug: 405361630
Test: btest CtsDevicePolicyTestCases
Change-Id: I1e6dc016bbf997a48274ac47e9d0d08c6b40ef74